strrchr working

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: rzamolo- <rzamolo-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/13 15:09:32 by rzamolo-          #+#    #+#             */
-/*   Updated: 2022/09/14 16:46:57 by rzamolo-         ###   ########.fr       */
+/*   Updated: 2022/09/20 12:55:52 by rzamolo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned long	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	unsigned long long	length;
